requirements.txt
amqp==2.5.2
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-celery-results==1.1.2
django-cors-headers==3.1.1
djangorestframework==3.10.3
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
numpy==1.17.4
psycopg2-binary==2.8.4
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
